refactor: colocate bulk upload with add media
