+++ frontend/src/entries/bulk-upload.js
@@ -1,5 +1,5 @@
 import '../static/css/tailwind.css';
 import { renderPage } from '../static/js/_helpers.js';
-import BulkUploadPage from '../features/bulk-upload/components/BulkUploadPage';
+import BulkUploadPage from '../features/add-media/bulk-upload/components/BulkUploadPage';
 
 renderPage('page-bulk-upload', BulkUploadPage);
