+++ frontend/src/features/add-media/bulk-upload/buildSubmitItems.js
