+++ frontend/src/features/add-media/bulk-upload/components/WizardFooter.test.jsx
