+++ frontend/src/features/add-media/bulk-upload/components/step1/UploadedFilesList.jsx
