+++ frontend/src/features/add-media/bulk-upload/components/step2/EnterDetails.jsx
@@ -1,5 +1,5 @@
-import { Text } from '../../../shared/components';
-import { EditorialPolicyNotice } from '../../../shared/components/upload-media';
+import { Text } from '../../../../shared/components';
+import { EditorialPolicyNotice } from '../../../../shared/components/upload-media';
 import useBulkUploadStore from '../../useBulkUploadStore';
 import { SubStepNav } from './SubStepNav';
 import { FileCard } from './FileCard';
